Env var RUN_AS controls the UID/GID of the running sbcl
The default value is 1000:1000 which should be replaced
with your own UID:GID on the Docker host. The directories

- /opt/quicklisp
- /cache

are chown'ed to that UID:GID, if they have another owner.

Note that this will take quite some time, so you should
always run the image with the same RUN_AS value.

gosu is used to drop root privileges and run sbcl:

https://github.com/tianon/gosu